% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joke.R
\name{joke}
\alias{joke}
\title{Randomly select and print joke}
\usage{
joke()
}
\value{
A randomly selected joke is printed and the text is invisibly returned
}
\description{
Prints a randomly selected Mitch Hedberg joke
}
