% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joke.R
\name{joke_on_startup}
\alias{joke_on_startup}
\title{Add mitchhedberg joke to startup}
\usage{
joke_on_startup()
}
\value{
This will append a local execution of mitchhedberg::joke() assuming
the session is interactive and the package has been installed.
}
\description{
Adds mitchhedberg::joke to user's .Rprofile, which results in printing a
joke at the start of interactive sessions
}
